refactor: plugins_materials
