+++ packages/plugins/materials/src/composable/block-compile.ts
