+++ packages/plugins/materials/src/composable/useResource.ts
